requirements.txt
# ===================================
# Core Framework
# ===================================
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1

# ===================================
# LangGraph Orchestration
# ===================================
langgraph==0.2.53
langchain>=0.3.13
langchain-core>=0.3.26
langchain-community>=0.3.13

# ===================================
# LLM Providers
# ===================================
google-generativeai==0.8.3
anthropic==0.39.0
groq==0.15.0
openai==1.60.0
langchain-google-genai==2.0.6
langchain-anthropic==0.3.3

# ===================================
# Redis & State Management
# ===================================
redis==5.2.1
hiredis==3.0.0  # High-performance Redis parser

# ===================================
# Forensics & Intelligence Extraction
# ===================================
tldextract==5.1.3  # Domain extraction and analysis
python-whois==0.9.4  # WHOIS lookup for domain age
phonenumbers==8.13.52  # Phone number parsing and validation
validators==0.34.0  # URL and email validation

# ===================================
# HTTP & Async
# ===================================
httpx==0.28.1  # Async HTTP client for callbacks
aiohttp==3.11.11
tenacity==9.0.0  # Retry mechanism with exponential backoff

# ===================================
# Logging & Monitoring
# ===================================
rich==13.9.4  # Beautiful terminal formatting
structlog==24.4.0  # Structured logging

# ===================================
# Testing
# ===================================
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0
faker==33.1.0  # Generate fake scam messages

# ===================================
# Development Tools
# ===================================
black==24.10.0
ruff==0.8.4
mypy==1.13.0
